+++ app/routes.py
@@ -218,6 +218,11 @@ def create_event():
         event.ended_at = pytz.utc.localize(datetime.utcnow())
         db.session.add(event)
     else:
+        entry_count = Event.query.filter_by(user_id=current_user.id).count()
+        if current_user.entry_limit <= entry_count:
+            oldest_event = Event.query.filter_by(user_id=current_user.id).order_by(Event.started_at.asc()).first()
+            db.session.delete(oldest_event)
+            flash('You have reached the {0} entry limit for your account. The oldest entry has been deleted.'.format(current_user.entry_limit), 'warning')
         new_event = Event(
             user_id=current_user.id,
             started_at=pytz.utc.localize(datetime.utcnow()))